site: added metadata for seo purposes
