+++ dash/app/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Gertrude for Parents</title>
     <meta name="description" content="Gertrude App Admin" />
-    <meta name="og:image" content="https://parents.gertrude.app/og-image.jpg" />
+    <meta name="og:image" content="https://parents.gertrude.app/og-images/main.jpg" />
     <link rel="stylesheet" type="text/css" href="/src/global.css" />
     <link rel="icon" type="image/png" href="/src/public/favicon.png" />
   </head>
