+++ site/app/docs/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-const DocsHomePage: React.FC = () => <div></div>;
+const DocsHomePage: React.FC = () => <div>redirecting...</div>;
 
 export default DocsHomePage;
