+++ site/components/articles/Sidebar.tsx
@@ -42,15 +42,15 @@ const Sidebar: React.FC<Props> = ({ open, setOpen }) => (
         </div>
       </div>
       <div className="flex flex-col gap-4">
-        <FancyLink
-          type="link"
-          href="https://parents.gertrude.app"
-          size="sm"
-          color="secondary"
-        >
+        <FancyLink type="link" href="https://parents.gertrude.app" size="sm">
           Log in
         </FancyLink>
-        <FancyLink type="link" href="https://parents.gertrude.app/signup" size="sm">
+        <FancyLink
+          type="link"
+          href="https://parents.gertrude.app/signup"
+          size="sm"
+          color="primary"
+        >
           Sign up
         </FancyLink>
       </div>
