+++ site/netlify.toml
@@ -4,9 +4,10 @@
   command = """
   cd ../../ && \
 	pnpm --filter site build && \
-  pnpm --filter site export && \
   cp site/_redirects site/out
   """
+
+# removed    pnpm --filter site export && \
 
 
 [[redirects]]
