+++ site/next.config.js
@@ -1,6 +1,7 @@
 /** @type {import('next').NextConfig} */
 
 const nextConfig = {
+  output: `export`,
   reactStrictMode: true,
   pageExtensions: [`tsx`, `md`],
   eslint: { ignoreDuringBuilds: true },
