+++ site/package.json
@@ -4,7 +4,6 @@
   "private": true,
   "scripts": {
     "start": "next dev",
-    "export": "next export",
     "build": "next build",
     "typecheck": "tsc --noEmit",
     "typecheck:watch": "tsc --noEmit --watch --preserveWatchOutput"
